invert priority.
Use lowest rather than highest priority when eliminating duplicates from
the epsilon closure.  This solves a few different issues: we were
artificially raising the priority of a ".*" prefix in order to get
leftmost longest, but this caused issues with sub-expressions that broke
things like ".*{a+b*a+}".  Instead, switch to lowest priority, and leave
the NFA numbered naturally.

Two subtle bugs caused that problem, though: 1) we were clearing
priorities after epsilon-closure (they definitely need to be retained
for later comparisons), and 2) when epsilon-closure generated new
states, it did so with a zero priority.  It needs the same (+ fs ts)
value as other introduced substates.

With these changes there's no need to raise the priority of a ".*"
prefix.

Also: dump-tdfa is now a single function that includes finals & nregs
in its output.