use list-of-ranges rather than set-of-ranges.
This roughly doubled the speed of the compiler when dealing with large machines (e.g. 5000 word OR).
Since the charset/merge operation took two sorted sets and returned a sorted set, the use of the set datatype was unnecessary and really slowed down charset operations, which had to be turned back into lists anyway.