Complete overhaul for Irken's new "3-way compare".
Irken switched to using 3-way compares in its container data structures.
This required changing a *lot* of code.

In order to pass (most of) the tdfa tests, I rolled back some of my
experiments trying to address the POSIX issue.  However, the one
troublesome test "a+b*a+" still fails.  [the fact that this test
passed in the past was a side-effect of a bug in properly managing
superstates].